requirements.txt
# Core Dependencies
prompt-toolkit>=3.0.43
rich>=13.7.0
google-generativeai>=0.3.0

# Additional utilities
python-dotenv>=1.0.0
colorama>=0.4.6

# Security
cryptography>=41.0.0

# Note: AWS CLI must be installed separately
# Install from: https://aws.amazon.com/cli/
